refactored to remove circular dependencies
