+++ apps/test/unit/lib/util/firehoseTest.js
@@ -1,0 +1,43 @@
+import {stub} from 'sinon';
+import {expect} from '../../../util/reconfiguredChai';
+import {validateFirehoseDataSize} from '@cdo/apps/lib/util/firehose';
+import logToCloud from '@cdo/apps/logToCloud';
+
+describe('firehoseDataSize', () => {
+  const maxDataJSONBytes = 65500;
+  const maxDataStringBytes = 4095;
+
+  beforeEach(() => {
+    stub(logToCloud, 'logError');
+  });
+
+  afterEach(() => {
+    logToCloud.logError.restore();
+  });
+
+  it('checks json size to send newrelic error', () => {
+    const valid_record = {data_json: 'x'.repeat(maxDataJSONBytes - 1)};
+    expect(validateFirehoseDataSize(valid_record)).not.to.be.true;
+    expect(logToCloud.logError).not.to.be.called;
+
+    const invalid_record = {data_json: 'x'.repeat(maxDataJSONBytes + 1)};
+    expect(() => {
+      validateFirehoseDataSize(invalid_record);
+    }).not.to.throw();
+    expect(logToCloud.logError).to.be.calledOnce;
+    expect(validateFirehoseDataSize(invalid_record)).to.be.true;
+  });
+
+  it('checks string size to send newrelic error', () => {
+    const valid_record = {data_string: 'x'.repeat(maxDataStringBytes - 1)};
+    expect(validateFirehoseDataSize(valid_record)).not.to.be.true;
+    expect(logToCloud.logError).not.to.be.called;
+
+    const invalid_record = {data_string: 'x'.repeat(maxDataStringBytes + 1)};
+    expect(() => {
+      validateFirehoseDataSize(invalid_record);
+    }).not.to.throw();
+    expect(logToCloud.logError).to.be.calledOnce;
+    expect(validateFirehoseDataSize(invalid_record)).to.be.true;
+  });
+});
